split userResolver to other file
